fix(core): worklet realm から TextEncoder を排除（snapshot codec を分割）
AudioWorkletGlobalScope は TextEncoder/TextDecoder を持たないため、snapshot.ts の module-scope `new TextEncoder()` が worklet module 評価時に throw し、registerProcessor 未到達で全 processor が browser で鳴らなくなっていた（node test は Node に global があり素通り）。snapshot.ts を worklet-safe な scalar codec / isPersistent と、main 専用の blob 文字列 codec（snapshotBlob.ts、TextEncoder 使用）に分割し、worklet バンドルから TextEncoder 参照を物理排除する。
